Added texture pack system
We can now load textures from a TOML file.  I have setup a temporary settings page which allows you to change the texture pack for backgrounds, card backs and cards.

Changed the scene system so that scenes behind a popup can be updated, and are aware they are behind a popup.