Added resizing support, updated settings page
Added layout resizing.  The card size is now calculated based on the window width and height.  I still have to rewrite a bit of the layout code.  Also fixed a bug where animating cards are not updated properly.

Updated the settings page so that all stored settings can be updated using the UI.  Also fixed the bug with reloading texture packs (was freeing a buffer that isn't malloc'd, and another buffer that was not initialized).

Misc changes / fixes:
There is no longer a max count for textures in a texture pack.  The buffer is resizable.
Fixed a bug where texture_from_path was still being passed 3 parameters.
Fixed a bug where a pack's folder was not unmounted in PHYSFS, leading to all subsequent pack loads being the same pack.
Updated printf's to be more standard.
Updated the Nuklear UI font to Roboto.