Complete web build
I have updated the web build with a new template, and fixed its compilation.

I have finished the card reveal animation, and the card preview in the settings menu.

Fixed a bug where the leaderboard page would crash if you had not won a game.

Started working more on the win animations.